# app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Trading API
  description: |
    FastAPI server for executing trading scripts and retrieving data files. 
    Serves as the backend for Sensylate, a React PWA for CSV-based portfolio analysis and strategy visualization.
    
    ## Features
    - Execute Python scripts with parameters (sync/async)
    - Real-time status updates via Server-Sent Events (SSE)
    - Retrieve CSV and JSON data files with security validation
    - Specialized portfolio update workflows
    - MCP server integration for AI assistant access
    
    ## Sensylate Integration
    This API handles the complete portfolio analysis workflow:
    1. Initial data loading via file listings
    2. CSV portfolio data retrieval and display
    3. Real-time portfolio updates with SSE streaming
    4. Automated trading signal detection
  version: 0.1.0
  contact:
    name: Trading API
    url: http://127.0.0.1:8000
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8000
    description: Development server
  - url: http://localhost:8000
    description: Local server
paths:
...
  /api/ma-cross/analyze:
    post:
      tags:
        - ma-cross
      summary: Parameter sensitivity for SMA and/or EMA strategies
      description: |
        Perform parameter sensitivty testing for Moving Average Crossover strategies on a portfolio of tickers. Supports both Simple Moving Average (SMA) and Exponential Moving Average (EMA).

        The analysis includes comprehensive backtesting with performance metrics such as:

        Total return and annualized return
        Sharpe and Sortino ratios
        Maximum drawdown
        Win rate and profit factor
        Trade statistics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MACrossRequest'
            examples:
              basic_analysis:
                summary: Basic MA Cross analysis
                value:
                  TICKER: ["AAPL", "MSFT"]
                  START_DATE: "2023-01-01"
                  END_DATE: "2023-12-31"
              advanced_analysis:
                summary: Advanced analysis with custom parameters
                value:
                  TICKER: ["BTC-USD", "ETH-USD", "SOL-USD"]
                  START_DATE: "2022-01-01"
                  END_DATE: "2023-12-31"
                  WINDOWS: 89
                  DIRECTION: "Long"
                  STRATEGY_TYPES: ["SMA", "EMA"]
                  USE_HOURLY: false
                  MIN_CRITERIA:
                    TRADES: 10
                    WIN_RATE: 0.5
                    PROFIT_FACTOR: 1.2
      responses:
        '200':
          description: Successful analysis response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MACrossResponse'
        '202':
          description: Asynchronous execution accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MACrossAsyncResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ma-cross/status/{execution_id}:
    get:
      tags:
        - ma-cross
      summary: Get analysis task status
      description: Check the status of an asynchronous MA Cross analysis task
      parameters:
        - name: execution_id
          in: path
          required: true
          description: Execution ID returned from async analysis request
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MACrossStatusResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/ma-cross/stream/{execution_id}:
    get:
      tags:
        - ma-cross
      summary: Stream analysis progress
      description: |
        Get real-time updates for MA Cross analysis via Server-Sent Events.
        The stream provides progress updates and partial results as the analysis runs.
      parameters:
        - name: execution_id
          in: path
          required: true
          description: Execution ID for the analysis to stream
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE stream with analysis progress updates.
                  
                  Example events:
                  ```
                  data: {"progress": 25, "message": "Fetching price data for AAPL"}
                  
                  data: {"progress": 50, "message": "Calculating indicators"}
                  
                  data: {"progress": 100, "message": "Analysis complete", "result": {...}}
                  ```
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MACrossRequest:
      type: object
      required:
        - TICKER
      properties:
        TICKER:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Trading symbol(s) to analyze
          example: ["AAPL", "MSFT", "GOOGL"]
        WINDOWS:
          type: integer
          minimum: 2
          maximum: 200
          default: 89
          description: Maximum window size for parameter analysis
        DIRECTION:
          type: string
          enum: ["Long", "Short"]
          default: "Long"
          description: Trading direction
        STRATEGY_TYPES:
          type: array
          items:
            type: string
            enum: ["SMA", "EMA"]
          default: ["SMA", "EMA"]
          description: Moving average types to analyze
        USE_HOURLY:
          type: boolean
          default: false
          description: Use hourly data instead of daily
        START_DATE:
          type: string
          format: date
          description: Start date for analysis (YYYY-MM-DD)
          example: "2023-01-01"
        END_DATE:
          type: string
          format: date
          description: End date for analysis (YYYY-MM-DD)
          example: "2023-12-31"
        MIN_CRITERIA:
          $ref: '#/components/schemas/MinimumCriteria'
        async_execution:
          type: boolean
          default: false
          description: Execute analysis asynchronously
    MinimumCriteria:
      type: object
      description: Minimum criteria for filtering portfolio results
      properties:
        TRADES:
          type: integer
          minimum: 0
          description: Minimum number of trades
        WIN_RATE:
          type: number
          minimum: 0
          maximum: 1
          description: Minimum win rate (0-1)
        EXPECTANCY_PER_TRADE:
          type: number
          description: Minimum expectancy per trade
        PROFIT_FACTOR:
          type: number
          minimum: 0
          description: Minimum profit factor
        SCORE:
          type: number
          description: Minimum portfolio score
        SORTINO_RATIO:
          type: number
          description: Minimum Sortino ratio
        BEATS_BNH:
          type: number
          description: Minimum percentage to beat Buy and Hold
    MACrossResponse:
      type: object
      required:
        - status
        - request_id
        - timestamp
        - ticker
        - strategy_types
        - portfolios
        - total_portfolios
        - filtered_portfolios
        - execution_time
      properties:
        status:
          type: string
          enum: ["success", "error"]
          description: Response status
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        ticker:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Analyzed ticker(s)
        strategy_types:
          type: array
          items:
            type: string
            enum: ["SMA", "EMA"]
          description: Strategy types analyzed
        portfolios:
          type: array
          items:
            $ref: '#/components/schemas/PortfolioMetrics'
          description: List of portfolio results
        total_portfolios:
          type: integer
          description: Total number of portfolios analyzed
        filtered_portfolios:
          type: integer
          description: Number of portfolios after filtering
        execution_time:
          type: number
          description: Analysis execution time in seconds
        request_summary:
          type: object
          additionalProperties: true
          description: Summary of request parameters
    PortfolioMetrics:
      type: object
      required:
        - ticker
        - strategy_type
        - short_window
        - long_window
        - total_return
        - annual_return
        - sharpe_ratio
        - sortino_ratio
        - max_drawdown
        - total_trades
        - winning_trades
        - losing_trades
        - win_rate
        - profit_factor
        - expectancy
        - score
        - beats_bnh
      properties:
        ticker:
          type: string
          description: Ticker symbol
        strategy_type:
          type: string
          enum: ["SMA", "EMA"]
          description: Moving average type
        short_window:
          type: integer
          description: Short MA period
        long_window:
          type: integer
          description: Long MA period
        total_return:
          type: number
          description: Total return percentage
        annual_return:
          type: number
          description: Annualized return percentage
        sharpe_ratio:
          type: number
          description: Sharpe ratio
        sortino_ratio:
          type: number
          description: Sortino ratio
        max_drawdown:
          type: number
          description: Maximum drawdown percentage
        total_trades:
          type: integer
          description: Total number of trades
        winning_trades:
          type: integer
          description: Number of winning trades
        losing_trades:
          type: integer
          description: Number of losing trades
        win_rate:
          type: number
          description: Win rate (0-1)
        profit_factor:
          type: number
          description: Profit factor
        expectancy:
          type: number
          description: Expectancy per trade
        score:
          type: number
          description: Overall portfolio score
        beats_bnh:
          type: number
          description: Percentage by which strategy beats Buy and Hold
        allocation_percent:
          type: number
          nullable: true
          description: Portfolio allocation percentage
        stop_loss_percent:
          type: number
          nullable: true
          description: Stop loss percentage
        position_size:
          type: number
          nullable: true
          description: Calculated position size
    MACrossAsyncResponse:
      type: object
      required:
        - status
        - execution_id
        - message
        - status_url
        - stream_url
        - timestamp
      properties:
        status:
          type: string
          enum: ["accepted"]
          description: Acceptance status
        execution_id:
          type: string
          format: uuid
          description: Task ID for tracking
        message:
          type: string
          description: Status message
        status_url:
          type: string
          format: uri
          description: URL to check task status
        stream_url:
          type: string
          format: uri
          description: URL for SSE stream
        timestamp:
          type: string
          format: date-time
          description: Request timestamp
    MACrossStatusResponse:
      type: object
      required:
        - execution_id
        - status
        - progress
      properties:
        execution_id:
          type: string
          format: uuid
          description: Execution identifier
        status:
          type: string
          enum: ["pending", "running", "completed", "failed"]
          description: Current task status
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Progress percentage
        message:
          type: string
          nullable: true
          description: Current status message
        result:
          type: object
          nullable: true
          description: Final result (when completed)
        error:
          type: string
          nullable: true
          description: Error message (if failed)
        start_time:
          type: string
          format: date-time
          description: Task start time
        end_time:
          type: string
          format: date-time
          nullable: true
          description: Task end time
...
    BaseResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [success, error, accepted]
          description: Status of the response
        message:
          type: string
          description: Optional message
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required:
            - message
          properties:
            status:
              type: string
              enum: [error]
              example: "error"
            message:
              type: string
              description: Error message
              example: "File not found"
            error_id:
              type: string
              description: Unique identifier for the error
              example: "a1b2c3d4e5f6"
            detail:
              type: string
              description: Detailed error information
              example: "The requested file does not exist or is not accessible"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future implementation)

# Global security (currently disabled for development)
# security:
#   - ApiKeyAuth: []
...