# app/api/routers/ma_cross.py
"""
MA Cross Router

This module provides API endpoints for MA Cross strategy analysis.
"""

from fastapi import APIRouter, HTTPException, Query, Path, Depends
from fastapi.responses import StreamingResponse
from typing import Dict, Any, List, Optional
import json
import asyncio
from datetime import datetime

from app.api.models.ma_cross import (
    MACrossRequest,
    MACrossResponse,
    MACrossAsyncResponse,
    MACrossStatusResponse,
    MACrossMetricsResponse
)
from app.api.models.response import ErrorResponse
from app.api.services.ma_cross_service import (
    MACrossService,
    MACrossServiceError
)
from app.api.utils.logging import setup_api_logging
from app.api.utils.cache import get_cache
from app.api.utils.validation import validate_ma_cross_request
from app.api.utils.middleware import rate_limit_analysis, rate_limit_cache
from app.api.utils.monitoring import get_metrics_collector
from app.api.utils.performance import get_concurrent_executor, get_request_optimizer

# Create router
router = APIRouter()

# Set up logging
log, _, logger, _ = setup_api_logging()

# Initialize service
ma_cross_service = MACrossService()

@router.post(
    "/analyze",
    response_model=MACrossResponse,
    responses={
        202: {"model": MACrossAsyncResponse, "description": "Asynchronous execution accepted"},
        400: {"model": ErrorResponse, "description": "Bad request"},
        429: {"description": "Rate limit exceeded"},
        500: {"model": ErrorResponse, "description": "Internal server error"}
    },
    summary="Parameter sensitivity analysis for MA Cross strategies",
    description="Perform parameter sensitivity testing for Moving Average Crossover strategies. "
                "Tests all window combinations from 2 to WINDOWS value to find optimal parameters. "
                "The analysis can be executed synchronously or asynchronously.",
    dependencies=[Depends(rate_limit_analysis)]
)
async def analyze_portfolio(request: MACrossRequest):
    """
    Execute MA Cross analysis on a portfolio.
    
    Args:
        request (MACrossRequest): MA Cross analysis request
        
    Returns:
        MACrossResponse or MACrossAsyncResponse: Analysis results or async execution info
        
    Raises:
        HTTPException: If the analysis fails
    """
    try:
        # Enhanced validation
        validation_error = validate_ma_cross_request(request)
        if validation_error:
            log(f"Request validation failed: {validation_error}")
            raise HTTPException(status_code=400, detail=validation_error)
        
        log(f"Analyzing portfolio with MA Cross strategy: {request.dict()}")
        
        # Execute analysis based on async flag
        if request.async_execution:
            response = ma_cross_service.analyze_portfolio_async(request)
            return response
        else:
            response = ma_cross_service.analyze_portfolio(request)
            return response
            
    except MACrossServiceError as e:
        log(f"MA Cross service error: {str(e)}", "error")
        raise HTTPException(status_code=500, detail=str(e))
    except ValueError as e:
        log(f"Invalid request: {str(e)}", "error")
        raise HTTPException(status_code=400, detail=str(e))
    except Exception as e:
        log(f"Unexpected error: {str(e)}", "error")
        raise HTTPException(status_code=500, detail=f"Unexpected error: {str(e)}")

@router.get(
    "/status/{execution_id}",
    response_model=MACrossStatusResponse,
    responses={
        404: {"model": ErrorResponse, "description": "Execution ID not found"},
        500: {"model": ErrorResponse, "description": "Internal server error"}
    },
    summary="Get MA Cross analysis status",
    description="Get the status of an asynchronous MA Cross analysis execution."
)
async def get_analysis_status(
    execution_id: str = Path(..., description="The execution ID returned from async analysis")
):
    """
    Get the status of an asynchronous MA Cross analysis.
    
    Args:
        execution_id (str): The execution ID
        
    Returns:
        MACrossStatusResponse: Current status of the analysis
        
    Raises:
        HTTPException: If the execution ID is not found
    """
    try:
        log(f"Getting status for execution ID: {execution_id}")
        
        status_info = ma_cross_service.get_task_status(execution_id)
        
        if not status_info:
            raise HTTPException(status_code=404, detail=f"Execution ID {execution_id} not found")
        
        return MACrossStatusResponse(**status_info)
        
    except HTTPException:
        raise
    except Exception as e:
        log(f"Error getting status: {str(e)}", "error")
        raise HTTPException(status_code=500, detail=f"Error getting status: {str(e)}")

@router.get(
    "/stream/{execution_id}",
    responses={
        200: {"description": "Server-sent event stream"},
        404: {"model": ErrorResponse, "description": "Execution ID not found"},
        500: {"model": ErrorResponse, "description": "Internal server error"}
    },
    summary="Stream MA Cross analysis updates",
    description="Stream real-time updates for an asynchronous MA Cross analysis using Server-Sent Events."
)
async def stream_analysis_updates(
    execution_id: str = Path(..., description="The execution ID returned from async analysis")
):
    """
    Stream real-time updates for an asynchronous MA Cross analysis.
    
    Args:
        execution_id (str): The execution ID
        
    Returns:
        StreamingResponse: Server-sent event stream
        
    Raises:
        HTTPException: If the execution ID is not found
    """
    async def generate_events():
        """Generate server-sent events for the analysis."""
        try:
            while True:
                # Get current status
                status_info = ma_cross_service.get_task_status(execution_id)
                
                if not status_info:
                    yield f"data: {json.dumps({'error': 'Execution ID not found'})}\n\n"
                    break
                
                # Send status update
                yield f"data: {json.dumps(status_info)}\n\n"
                
                # If analysis is complete or failed, stop streaming
                if status_info['status'] in ['completed', 'failed']:
                    break
                
                # Wait before next update
                await asyncio.sleep(1)
                
        except Exception as e:
            yield f"data: {json.dumps({'error': str(e)})}\n\n"
    
    try:
        # Check if execution ID exists
        status_info = ma_cross_service.get_task_status(execution_id)
        if not status_info:
            raise HTTPException(status_code=404, detail=f"Execution ID {execution_id} not found")
        
        return StreamingResponse(
            generate_events(),
            media_type="text/event-stream",
            headers={
                "Cache-Control": "no-cache",
                "Connection": "keep-alive"
            }
        )
        
    except HTTPException:
        raise
    except Exception as e:
        log(f"Error streaming updates: {str(e)}", "error")
        raise HTTPException(status_code=500, detail=f"Error streaming updates: {str(e)}")

@router.get(
    "/metrics",
    response_model=MACrossMetricsResponse,
    responses={
        500: {"model": ErrorResponse, "description": "Internal server error"}
    },
    summary="Get available MA Cross metrics",
    description="Get information about the metrics calculated by the MA Cross strategy."
)
async def get_available_metrics():
    """
    Get information about available MA Cross metrics.
    
    Returns:
        MACrossMetricsResponse: Available metrics information
    """
    try:
        metrics_info = {
            "available_metrics": [
                {
                    "name": "total_return",
                    "description": "Total return percentage over the analysis period",
                    "unit": "percentage"
                },
                {
                    "name": "sharpe_ratio",
                    "description": "Risk-adjusted return metric",
                    "unit": "ratio"
                },
                {
                    "name": "max_drawdown",
                    "description": "Maximum peak-to-trough decline",
                    "unit": "percentage"
                },
                {
                    "name": "win_rate",
                    "description": "Percentage of winning trades",
                    "unit": "percentage"
                },
                {
                    "name": "profit_factor",
                    "description": "Ratio of gross profit to gross loss",
                    "unit": "ratio"
                },
                {
                    "name": "trades_count",
                    "description": "Total number of trades executed",
                    "unit": "count"
                },
                {
                    "name": "avg_trade_duration",
                    "description": "Average duration of trades in days",
                    "unit": "days"
                },
                {
                    "name": "calmar_ratio",
                    "description": "Return over maximum drawdown",
                    "unit": "ratio"
                }
            ],
            "metric_categories": {
                "performance": ["total_return", "profit_factor"],
                "risk": ["max_drawdown", "sharpe_ratio", "calmar_ratio"],
                "execution": ["trades_count", "win_rate", "avg_trade_duration"]
            }
        }
        
        return MACrossMetricsResponse(**metrics_info)
        
    except Exception as e:
        log(f"Error getting metrics info: {str(e)}", "error")
        raise HTTPException(status_code=500, detail=f"Error getting metrics info: {str(e)}")

@router.get(
    "/health",
    responses={
        200: {"description": "Service is healthy"},
        503: {"model": ErrorResponse, "description": "Service unavailable"}
    },
    summary="MA Cross service health check",
    description="Check if the MA Cross service is healthy and ready to process requests."
)
async def health_check():
    """
    Check MA Cross service health.
    
    Returns:
        dict: Health status
    """
    try:
        # TODO: Add more comprehensive health checks
        # - Check if scanner module is available
        # - Check if data directories are accessible
        # - Check thread pool status
        
        return {
            "status": "healthy",
            "service": "ma_cross",
            "timestamp": datetime.utcnow().isoformat()
        }
        
    except Exception as e:
        log(f"Health check failed: {str(e)}", "error")
        raise HTTPException(status_code=503, detail=f"Service unhealthy: {str(e)}")


@router.get(
    "/cache/stats",
    summary="Get cache statistics",
    description="Returns cache performance metrics including hit rate and size",
    dependencies=[Depends(rate_limit_cache)]
)
async def get_cache_stats():
    """Get cache statistics."""
    try:
        cache = get_cache()
        stats = cache.get_stats()
        return {
            "status": "success",
            "cache_stats": stats,
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to get cache stats: {str(e)}")


@router.post(
    "/cache/invalidate",
    summary="Invalidate cache entries",
    description="Clear cache entries, optionally filtered by ticker symbol",
    dependencies=[Depends(rate_limit_cache)]
)
async def invalidate_cache(
    ticker: Optional[str] = Query(None, description="Ticker symbol to invalidate (clears all if not provided)")
):
    """Invalidate cache entries."""
    try:
        cache = get_cache()
        
        if ticker:
            invalidated = cache.invalidate_pattern(ticker)
            message = f"Invalidated {invalidated} cache entries for ticker: {ticker}"
        else:
            cache.invalidate_all()
            message = "Invalidated all cache entries"
        
        return {
            "status": "success",
            "message": message,
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to invalidate cache: {str(e)}")


@router.post(
    "/cache/cleanup",
    summary="Clean up expired cache entries",
    description="Remove expired entries from cache and return cleanup statistics",
    dependencies=[Depends(rate_limit_cache)]
)
async def cleanup_cache():
    """Clean up expired cache entries."""
    try:
        cache = get_cache()
        cleanup_stats = cache.cleanup()
        
        return {
            "status": "success",
            "cleanup_stats": cleanup_stats,
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to cleanup cache: {str(e)}")


@router.get(
    "/metrics",
    summary="Get API metrics",
    description="Returns API performance and usage metrics"
)
async def get_api_metrics(
    hours: int = Query(1, ge=1, le=24, description="Number of hours to include in metrics")
):
    """Get API performance metrics."""
    try:
        metrics = get_metrics_collector()
        request_stats = metrics.get_request_stats(hours=hours)
        system_stats = metrics.get_system_stats()
        
        return {
            "status": "success",
            "metrics": {
                "requests": request_stats,
                "system": system_stats,
                "period_hours": hours
            },
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to get metrics: {str(e)}")


@router.get(
    "/health/detailed",
    summary="Get detailed health status",
    description="Returns comprehensive health information including system metrics"
)
async def get_detailed_health():
    """Get detailed service health status."""
    try:
        metrics = get_metrics_collector()
        health_status = metrics.get_health_status()
        
        return {
            "status": "success",
            "health": health_status,
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to get health status: {str(e)}")


@router.post(
    "/metrics/cleanup",
    summary="Clean up old metrics",
    description="Remove old metrics data to free memory"
)
async def cleanup_metrics():
    """Clean up old metrics data."""
    try:
        metrics = get_metrics_collector()
        cleanup_stats = metrics.cleanup_old_metrics()
        
        return {
            "status": "success",
            "cleanup_stats": cleanup_stats,
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to cleanup metrics: {str(e)}")


@router.get(
    "/performance/stats",
    summary="Get performance statistics",
    description="Returns performance optimization and timing statistics"
)
async def get_performance_stats():
    """Get performance statistics."""
    try:
        executor = get_concurrent_executor()
        optimizer = get_request_optimizer()
        
        executor_stats = executor.get_stats()
        timing_stats = optimizer.get_timing_stats()
        
        return {
            "status": "success",
            "performance": {
                "executor": executor_stats,
                "timing": timing_stats
            },
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to get performance stats: {str(e)}")


@router.get(
    "/rate-limit/stats",
    summary="Get rate limiting statistics",
    description="Returns current rate limiting status for all clients"
)
async def get_rate_limit_stats():
    """Get rate limiting statistics."""
    try:
        from app.api.utils.rate_limiter import get_analysis_limiter, get_cache_limiter
        
        analysis_limiter = get_analysis_limiter()
        cache_limiter = get_cache_limiter()
        
        return {
            "status": "success",
            "rate_limits": {
                "analysis": analysis_limiter.get_stats(),
                "cache": cache_limiter.get_stats()
            },
            "timestamp": datetime.utcnow().isoformat()
        }
    except Exception as e:
        raise HTTPException(status_code=500, detail=f"Failed to get rate limit stats: {str(e)}")